made a bunch of changes, solved some exercises, changed cmakelists, etc

+++ src/divisibleBy.c
@@ -2,12 +2,12 @@
 #include <stdio.h>
 
 // Problem 3.1 - Given an interval, return all numbers within that interval that when divided by X are equal to Y
-DynArr divisibleBy(unsigned int start, unsigned int end, unsigned int x, unsigned int y) {
+DynArr divisibleBy(uint32_t start, uint32_t end, uint32_t x, uint32_t y) {
 	DynArr result;
 	result.size = 0;
 	result.arr = malloc(result.size * sizeof(int));
 
-	for (unsigned int i = start; i <= end; i++) {
+	for (size_t i = start; i <= end; i++) {
 		if (i % x == y) {
 			result.arr[result.size] = i;
 			result.size++;
